pyproject.toml
[tool.poetry]
name = "line-chatgpt"

[tool.poetry.dependencies]
python = "^3.9"
line-bot-sdk = "^2.4.2"
openai = "^0.27.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
